+++ 1-node-farm/index.js
@@ -1,7 +1,7 @@
 const http = require("http");
 const url = require("url");
 const fs = require("fs");
-
+const replaceTemplate = require("./modules/replaceTemplate");
 /////////////////////////////////
 // SERVER
 const tempOverview = fs.readFileSync(
@@ -12,7 +12,6 @@ const tempCard = fs.readFileSync(
   `${__dirname}/templates/card.html`,
   "utf-8",
 );
-
 const tempProduct = fs.readFileSync(
   `${__dirname}/templates/product.html`,
   "utf-8",
@@ -20,47 +19,32 @@ const tempProduct = fs.readFileSync(
 const data = fs.readFileSync(`${__dirname}/dev-data/data.json`, "utf-8");
 const dataObj = JSON.parse(data);
 
-const replaceTemplate = (temp, product) => {
-  ///{%PRODUCTNAME%}/g => g means global to replace all match REGEX
-  let output = temp.replace(/{%PRODUCTNAME%}/g, product.productName);
-  output = output.replace(/{%IMAGE%}/g, product.image);
-  output = output.replace(/{%PRICE%}/g, product.price);
-  output = output.replace(/{%FROM%}/g, product.from);
-  output = output.replace(/{%NUTRIENTS%}/g, product.nutrients);
-  output = output.replace(/{%QUANTITY%}/g, product.quantity);
-  output = output.replace(/{%DESCRIPTION%}/g, product.description);
-  output = output.replace(/{%ID%}/g, product.id);
-
-  //If Product is NOT Organic
-  //Replacing the placeholder /{%NOT_ORGANIC%}/ with Class Name: "not-organic"
-  if (!product.organic) {
-    output = output.replace(/{%NOT_ORGANIC%}/g, "not-organic");
-  }
-  return output;
-};
-
 const server = http.createServer((req, res) => {
-  //const { query, pathname } = url.parse(req.url, true);
-  const pathname = req.url;
+  //url.parse => to parse req.url to become url object contains query, pathname & many more
+  //=> To parse query, say /product?id=0,  into object {query: {id: '0'}}
+  const { query, pathname } = url.parse(req.url, true);
+
   // Overview page
   if (pathname === "/" || pathname === "/overview") {
     res.writeHead(200, {
       "Content-type": "text/html",
     });
-
-    //Replace the Place Holders in Card HTML with data in dataObj
-    const cardsHTML = dataObj.map((el) => replaceTemplate(tempCard, el)).join(
+    const cardsHtml = dataObj.map((el) => replaceTemplate(tempCard, el)).join(
       "",
     );
-    //Replace "{PRODUCT_CARDS}" with cardsHTML template
-    const output = tempOverview.replace("{%PRODUCT_CARDS%}", cardsHTML);
-
+    const output = tempOverview.replace("{%PRODUCT_CARDS%}", cardsHtml);
     res.end(output);
   } else if (pathname === "/product") {
-    res.end("This is the PRODUCT");
+    res.writeHead(200, {
+      "Content-type": "text/html",
+    });
+    const product = dataObj[query.id];
+
+    const output = replaceTemplate(tempProduct, product);
+    res.end(output);
   } else if (pathname === "/api") {
     res.writeHead(200, {
-      "Content-type": "application/json",
+      "Content-type": "application/json", //Res to browser JSON Data
     });
     res.end(data);
   } else {
@@ -73,6 +57,6 @@ const server = http.createServer((req, res) => {
   }
 });
 
-server.listen(8000, "localhost", () => {
+server.listen(8000, "127.0.0.1", () => {
   console.log("Listening to requests on port 8000");
 });
